+++ resources/fsUtils.ts
@@ -4,6 +4,7 @@ const fs = require("fs");
 
 export function deleteAll() {
   const OUTPUT_FOLDER = params.OUTPUT_DIR;
+  const START_PAGE = params.START_PAGE;
   const rootPath = path.resolve(__dirname, "../../");
   const outputFolder = path.join(rootPath, OUTPUT_FOLDER);
   const pagesFolder = path.join(outputFolder, "/pages");
@@ -13,23 +14,31 @@ export function deleteAll() {
     const pagesFiles = fs.readdirSync(pagesFolder);
     const compiledFiles = fs.readdirSync(outputFolder);
 
-    // Loop through the files and delete each one
+    // Delete files in the pages folder
     for (const file of pagesFiles) {
       const filePath = path.join(pagesFolder, file);
+      const ext = path.extname(filePath);
+      const filename = path.basename(filePath, ext);
 
-      // Check if it's a file (not a subdirectory)
       if (fs.statSync(filePath).isFile()) {
+        const pageCount = +filename.split("-")[2];
+
+        if (pageCount < START_PAGE) {
+          console.log(`Skipped deleting file: ${filename}`);
+          continue;
+        }
+
         fs.unlinkSync(filePath); // Delete the file synchronously
         console.log(`Deleted file: ${filePath}`);
       }
     }
 
+    // delete compiled
     for (const file of compiledFiles) {
       const filePath = path.join(outputFolder, file);
 
-      // Check if it's a file (not a subdirectory)
       if (fs.statSync(filePath).isFile()) {
-        fs.unlinkSync(filePath); // Delete the file synchronously
+        fs.unlinkSync(filePath);
         console.log(`Deleted file: ${filePath}`);
       }
     }
